Don't force Int64, for 32-bit platforms

+++ src/CVaRRiskParity.jl
@@ -15,7 +15,7 @@ module CVaRRiskParity
 include("cvar_cp.jl") # Cutting Plane algorithm, and related utilities
 
 """
-    cvar_rbp(B::Vector{Float64}, alpha::Float64, rel_losses::Array{Float64,2}; tol::Float64=1e-6, maxiters::Int64=1000)
+    cvar_rbp(B::Vector{Float64}, alpha::Float64, rel_losses::Array{Float64,2}; tol::Float64=1e-6, maxiters::Int=1000)
 
 Compute the investment weights on the assets in order to build a
 CV@R-`alpha` risk budgeting portfolio given risk appetites in `B`
@@ -29,7 +29,7 @@ Returns (failed, w) :: Bool, Vector{Float64}
 
 For more details on the algorithm, see `cutting_planes`.
 """
-function cvar_rbp(B::Vector{Float64}, alpha::Float64, rel_losses::Array{Float64,2}; tol::Float64=1e-6, maxiters::Int64=1000)
+function cvar_rbp(B::Vector{Float64}, alpha::Float64, rel_losses::Array{Float64,2}; tol::Float64=1e-6, maxiters::Int=1000)
   f, w, t = cutting_planes(B, alpha, rel_losses; tol=tol, maxiters=maxiters, debug=0)
   return f == 1, w
 end
